Refactor radioActions
